# lib/api-zod/src/generated/types/trade.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeExchange } from './tradeExchange';
import type { TradeStatus } from './tradeStatus';

export interface Trade {
  id: number;
  symbol: string;
  exchange: TradeExchange;
  quantity: number;
  entryPrice: number;
  stopPrice?: number | null;
  targetPrice?: number | null;
  plannedExitAt?: Date | null;
  status: TradeStatus;
  exitPrice?: number | null;
  exitedAt?: Date | null;
  exitReason?: string | null;
  followUpStatus: string;
  followUpNotes?: string | null;
  thesis?: string | null;
  createdAt: Date;
  updatedAt: Date;
}