# lib/api-zod/src/generated/types/tradeRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeRequestExchange } from './tradeRequestExchange';

export interface TradeRequest {
  /** @minLength 1 */
  symbol: string;
  exchange?: TradeRequestExchange;
  /** @minimum 1 */
  quantity: number;
  /** @exclusiveMinimum 0 */
  entryPrice: number;
  /** @exclusiveMinimum 0 */
  stopPrice?: number;
  /** @exclusiveMinimum 0 */
  targetPrice?: number;
  plannedExitAt?: Date;
  thesis?: string;
}