# lib/api-zod/src/generated/types/tradeUpdateRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeUpdateRequestStatus } from './tradeUpdateRequestStatus';

export interface TradeUpdateRequest {
  status?: TradeUpdateRequestStatus;
  /** @exclusiveMinimum 0 */
  exitPrice?: number;
  exitedAt?: Date;
  exitReason?: string;
  followUpStatus?: string;
  followUpNotes?: string;
}